# src/api/openapi.yaml
openapi: 3.0.0
info:
  title: ICP Analyzer API
  version: 1.0.0
  description: API for analyzing customer testimonials from websites
  contact:
    name: API Support
    email: support@icpanalyzer.com

servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://api.icpanalyzer.com
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AnalysisRequest:
      type: object
      required:
        - urls
        - analysis_type
      properties:
        urls:
          type: array
          items:
            type: string
          description: List of website URLs to analyze
        analysis_type:
          type: string
          enum: [single, comparative, competitive]
          description: Type of analysis to perform
        days_back:
          type: integer
          default: 30
          description: Number of days of historical data for competitive analysis
        include_advanced:
          type: boolean
          default: false
          description: Whether to include advanced analytics

    ReportRequest:
      type: object
      required:
        - analysis_data
        - report_type
      properties:
        analysis_data:
          type: object
          description: Analysis results to include in the report
        report_type:
          type: string
          enum: [pdf, pptx, docx, excel]
          description: Format of the report to generate
        metrics:
          type: array
          items:
            type: string
          description: Specific metrics to include in the report
        branding:
          type: object
          properties:
            company_name:
              type: string
              description: Company name for report branding

    WebhookRequest:
      type: object
      required:
        - webhook_url
      properties:
        webhook_url:
          type: string
          description: URL to receive analysis completion notifications

    ExportRequest:
      type: object
      required:
        - platform
        - analysis_data
      properties:
        platform:
          type: string
          enum: [salesforce, hubspot]
          description: Platform to export data to
        analysis_data:
          type: object
          description: Analysis results to export

paths:
  /api/analyze:
    post:
      summary: Analyze website testimonials
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequest'
      responses:
        '200':
          description: Analysis completed successfully
        '400':
          description: Invalid request parameters
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /api/report:
    post:
      summary: Generate analysis report
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '200':
          description: Report generated successfully
        '400':
          description: Invalid request parameters
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /api/webhook:
    post:
      summary: Register webhook for analysis notifications
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookRequest'
      responses:
        '200':
          description: Webhook registered successfully
        '400':
          description: Invalid request parameters
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /api/export:
    post:
      summary: Export analysis results to external platform
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Export completed successfully
        '400':
          description: Invalid request parameters
        '401':
          description: Unauthorized
        '500':
          description: Server error 